requirements.txt
appnope==0.1.0
certifi==2017.4.17
chardet==3.0.3
cycler==0.10.0
decorator==4.0.11
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jupyter-client==5.0.1
jupyter-core==4.3.0
matplotlib==2.0.2
networkx==1.11
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.7
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
pytesseract==0.1.7
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
pyzmq==16.0.2
reportlab==3.4.0
requests==2.17.3
requests-file==1.4.2
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
tldextract==2.1.0
tornado==4.5.1
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
